Rearrange and use frost simulate keygen
